datasheets

+++ src/Hajir.Crm/Sales/SaleQuoteLine.cs
@@ -17,7 +17,7 @@ namespace Hajir.Crm.Sales
         public string Name { get; set; }
         public decimal? Discount { get; set; }
         public int? PercentDiscount { get; set; }
-        public long? GuaranteeMonth { get; set; }
+        public int? GuaranteeMonth { get; set; }
         public decimal? ExtendedAmount { get; set; }
         public decimal? BaseAmount { get; set; }
         public decimal? Tax { get; set; }
